% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2hex2.R
\name{col2hex2}
\alias{col2hex2}
\title{Output hex code for built-in R color name}
\usage{
col2hex2(cname)
}
\arguments{
\item{cname}{character string name of a built-in r color}
}
\value{
character string with hex code
}
\description{
Built on the col2hex() function from the gplot package by Tal Galili with added error handling
}
\examples{
col2hex("firebrick4")
}
