% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2htmlspan_notext.R
\name{col2htmlspan_notext}
\alias{col2htmlspan_notext}
\title{Generate HTML to display background color for a built-in R color}
\usage{
col2htmlspan_notext(colorname, character_length = 30)
}
\arguments{
\item{colorname}{character string of a built-in R color name}

\item{text}{character string for text you want to appear over the background color, if any}

\item{textcolor}{character string for the color of the text you want to appear over the background color, if any}
}
\value{
character string with HTML
}
\description{
Generate HTML to display background color for a built-in R color
}
\examples{
col2htmlspan_notext("firebrick4")
}
