% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\name{col2rgb2}
\alias{col2rgb2}
\title{Get named integer of rgb from R built-in color name}
\usage{
col2rgb2(colorname)
}
\arguments{
\item{colorname}{character string name of built-in R color}
}
\value{
named integers
}
\description{
Get named integer of rgb from R built-in color name
}
\examples{
col2rgb2("firebrick4")
}
