% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\name{nearRcolor}
\alias{nearRcolor}
\title{Find close R colors() to a given color (original by Marius Hofert)
using Euclidean norm in (HSV / RGB / ...) color space}
\usage{
nearRcolor(
  rgb,
  cSpace = c("hsv", "rgb255", "Luv", "Lab"),
  dist = switch(cSpace, hsv = 0.1, rgb255 = 30, Luv = 15, Lab = 12)
)
}
\arguments{
\item{rgb}{Can be character string of built-in R color name like "tomato" or color rgb as vector ex c(255,99,71)}
}
\description{
Find close R colors() to a given color (original by Marius Hofert)
using Euclidean norm in (HSV / RGB / ...) color space
}
\author{
Marius Hofert, originally
}
