% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\name{plotCol}
\alias{plotCol}
\title{Compare Colors In A Plot}
\usage{
plotCol(col, nrow = 1, ncol = ceiling(length(col)/nrow), txt.col = "black")
}
\arguments{
\item{col}{the color name}

\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{txt.col}{text color}
}
\value{
the grid layout, invisibly
}
\description{
Compare Colors In A Plot
}
\author{
Marius Hofert, originally
}
