% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert rgb color value to hex code}
\usage{
rgb2hex(r, g, b)
}
\arguments{
\item{r}{integer r value}

\item{g}{integer g value}

\item{b}{integer b value}
}
\value{
character string
}
\description{
Convert rgb color value to hex code
}
\examples{
rgb2hex(255,0,0)
}
\author{
Hadley Wickham
}
