% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_color_table.R
\name{create_color_table}
\alias{create_color_table}
\title{Generate an interactive, searchable table of built-in R colors}
\usage{
create_color_table(page_length = 25)
}
\arguments{
\item{page_length}{integer for length of table page, defaults to 25}
}
\value{
DT object with HTML of table that displays correctly in RStudio
}
\description{
Generate an interactive, searchable table of built-in R colors
}
